+++ shift-scan/src/app/(routes)/dashboard/clock-out/(components)/clock-out-Verification/laborClockOut.tsx
@@ -1,4 +1,8 @@
-import { useScanData } from "@/app/context/JobSiteScanDataContext";
+import {
+  RemoveCookiesAtClockOut,
+  setStartingMileage,
+} from "@/actions/cookieActions";
+import { updateTimeSheet } from "@/actions/timeSheetActions";
 import { Bases } from "@/components/(reusable)/bases";
 import { Buttons } from "@/components/(reusable)/buttons";
 import { Contents } from "@/components/(reusable)/contents";
@@ -10,25 +14,48 @@ import { Labels } from "@/components/(reusable)/labels";
 import { Titles } from "@/components/(reusable)/titles";
 import { Clock } from "@/components/clock";
 import { useTranslations } from "next-intl";
+import { useRouter } from "next/navigation";
 import { useState } from "react";
 
 export const LaborClockOut = ({
-  handleButtonClick,
   scanResult,
   savedCostCode,
-  formRef,
-  isSubmitting,
   prevStep,
 }: {
-  handleButtonClick: () => void;
   scanResult: string | undefined;
   savedCostCode: string | null;
-  formRef: React.RefObject<HTMLFormElement>;
-  isSubmitting: boolean;
   prevStep: () => void;
 }) => {
   const t = useTranslations("ClockOut");
   const [date] = useState(new Date());
+  const router = useRouter();
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault(); // Prevents default form submission
+
+    try {
+      let timeSheetId = null;
+      const response = await fetch("/api/getRecentTimecard");
+      const tsId = await response.json();
+      timeSheetId = tsId.id;
+
+      if (!timeSheetId) {
+        alert("No valid TimeSheet ID was found. Please try again later.");
+        return;
+      }
+
+      // Collect form data
+      const formData = new FormData(event.currentTarget);
+      formData.append("id", timeSheetId); // Ensure TimeSheet ID is included
+
+      await updateTimeSheet(formData);
+      localStorage.clear();
+      setStartingMileage("");
+      RemoveCookiesAtClockOut();
+      router.push("/");
+    } catch (error) {
+      console.error("Failed to submit the time sheet:", error);
+    }
+  };
 
   return (
     <Bases>
@@ -131,7 +158,7 @@ export const LaborClockOut = ({
                     background={"darkBlue"}
                     className="h-full w-[100%] sm:w-[90%] md:w-[90%] lg:w-[80%] xl:w-[80%] 2xl:w-[80%]  border-[3px]   border-black p-8 "
                   >
-                    <form ref={formRef} className="h-full">
+                    <form onClick={handleSubmit} className="h-full">
                       <Inputs
                         type="hidden"
                         name="endTime"
@@ -146,7 +173,7 @@ export const LaborClockOut = ({
                       />
                       {/* Cancel out the button shadow with none background  and then add a class name */}
                       <Buttons
-                        onClick={handleButtonClick}
+                        type="submit"
                         className="bg-app-green flex justify-center items-center p-4 rounded-[10px] text-black font-bold"
                       >
                         <Clock time={date.getTime()} />
